# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Staging Superapp API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server (Docker)"
        }
    ],
    "paths": {
...
        "/api/v1/sales/returns": {
            "get": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Get list of sales returns",
                "operationId": "e7e9321f83b763526b334561f7de0ffc",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[source]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Create a sales return (with or without order/invoice)",
                "operationId": "79b24388c7331545da1a29165cce88bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_id",
                                    "created_by",
                                    "items"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "location_id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "manual",
                                        "enum": [
                                            "manual",
                                            "marketplace"
                                        ]
                                    },
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "reason": {
                                        "type": "string"
                                    },
                                    "created_by": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "qty"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "condition": {
                                                    "type": "string",
                                                    "enum": [
                                                        "GOOD",
                                                        "DAMAGE"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sales return berhasil dibuat"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/returns/unprocessed": {
            "get": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Get unprocessed marketplace returns",
                "operationId": "1c7b2d3e717f001c95762dd8b1de3e39",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/returns/{id}": {
            "get": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Get sales return details",
                "operationId": "b6e8d0f51b77218db6f62c1a8a1e0873",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Return tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/returns/{id}/accept": {
            "post": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Accept a return (stock masuk ke gudang)",
                "operationId": "e49b8df221b19b3cc64925891277d37d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "processed_by"
                                ],
                                "properties": {
                                    "processed_by": {
                                        "type": "string",
                                        "example": "warehouse_staff"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return diterima, Inbound GRN dibuat"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/returns/{id}/reject": {
            "post": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Reject a return (stock tidak berubah)",
                "operationId": "5484e55ced87bb13bb47166ac52177c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "processed_by"
                                ],
                                "properties": {
                                    "processed_by": {
                                        "type": "string",
                                        "example": "warehouse_staff"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return ditolak"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/returns/{id}/complete": {
            "post": {
                "tags": [
                    "Sales Returns"
                ],
                "summary": "Mark return as complete",
                "operationId": "e12dab321350b47819bc0e1ca3e41cad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "processed_by"
                                ],
                                "properties": {
                                    "processed_by": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return selesai"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}